feat(cnpg-system): remove cnpg
cnpg is cool (we use it at work), but it is not ideal for my homelab
usecase. I only have one replica each, so no great improvement with HA.
backups would also require either an s3 compatible storage or a
snapshotter, both of which I don't have. so for now, I'll just stay with
the bitnami chart ¯\_(ツ)_/¯